requirements.txt
caldav>=1.3
openpyxl>=3.1
beautifulsoup4>=4.12
requests>=2.31
pytz>=2024.1
python-dateutil>=2.9
holidays>=0.60
